backend/requirements.txt
fastapi
uvicorn
faiss-cpu
requests
numpy
sentence-transformers